+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+### 1.0.0
+
+##### 22/08/2021
+
+- Named exports
+
+```js
+import { getCountryName } from "@brixtol/country-names";
+```
+
 ### 0.1.0
 
 ##### 17/08/2021